+++ e2j2/helpers/__init__.py
@@ -1,7 +1,6 @@
-from e2j2.helpers import constants, parsers, templates
+from e2j2.helpers import constants, templates
 
 __all__ = [
     constants,
-    parsers,
     templates
 ]