+++ e2j2/tags/list.py
@@ -1,0 +1,5 @@
+import re
+
+
+def parse(value):
+    return re.split(r",\s*", value)