+++ e2j2/tags/vault.py
@@ -1,0 +1,2 @@
+def parse(vault_key):
+    pass